requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
requests==2.32.5
toons==0.5.3
urllib3==2.6.3
prompt_toolkit==3.0.47
pyyaml==6.0.3
faiss-cpu==1.9.0
numpy==2.2.3
